refactor: 고객페이지 컴포넌트 계층 분리및 커스텀 훅 생성
